Feat: 이미지 파일 assets 이동, Header, Footer 분리, landig->signup 화면 전환 구현 (#6)
