+++ src/app/routes.const.ts
@@ -1,4 +1,5 @@
 export const ROUTES = {
   LANDING: "/",
   LOGIN: "/login",
+  SIGNUP: "/signup",
 } as const;
