+++ src/assets/hero-image.png
