--- src/components/Layout.tsx
@@ -1,12 +1,0 @@
-// src/components/Layout.tsx
-import { Outlet } from "react-router-dom";
-
-const Layout = () => {
-  return (
-    <div>
-      <Outlet />
-    </div>
-  );
-};
-
-export default Layout;
