+++ src/pages/landing/LandingPage.tsx
@@ -1,0 +1,13 @@
+import FeatureSection from "./components/FeatureSection";
+import HeroSection from "./components/HeroSection";
+import LandingHeader from "./components/LandingHeader";
+
+export default function LandingPage() {
+  return (
+    <>
+      <LandingHeader />
+      <HeroSection />
+      <FeatureSection />
+    </>
+  );
+}
