+++ src/pages/signup/SignupPage.tsx
@@ -1,0 +1,3 @@
+export default function SignupPage() {
+  return <div>Signup Page</div>;
+}
