+++ src/styles/layout.css
@@ -9,6 +9,16 @@
   padding: var(--space-7) 0;
 }
 
+.layout {
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+}
+
+.layout__main {
+  flex: 1;
+}
+
 @media (max-width: 768px) {
   .container {
     padding: 0 16px;
